Porting TEMPLATE calls to ParserPipeline.
