+++ SciXParser/API/avro_serializer.py
