+++ SciXParser/tests/parser/base.py
