--- SciXTEMPLATE/AVRO_schemas/TEMPLATEOutputSchema.avsc
@@ -1,25 +1,0 @@
-{
-  "type": "record",
-  "name": "TemplateOutput",
-  "fields" : [
-    {
-        "name": "record_id",
-        "type": ["string", "null"],
-        "doc": "The job ID generated by storing it in postgres."
-    },
-    {
-        "name": "record_data",
-        "doc": "The data for a given record.",
-        "type": ["string", "null"]
-    },
-    {
-        "name": "task",
-        "doc": "An enumerated type for specifying the data source.",
-        "type": {
-            "name": "taskenum",
-            "type": "enum",
-            "symbols" : ["SYMBOL1", "SYMBOL2", "SYMBOL3", "SYMBOL4"]
-        }
-    }
-  ]
-}
