--- SciXTEMPLATE/alembic/versions/6bfebb4af6ba_initialize_database.py
@@ -1,40 +1,0 @@
-"""initialize_database
-
-Revision ID: 6bfebb4af6ba
-Revises:
-Create Date: 2022-11-02 11:58:51.077956
-
-"""
-import sqlalchemy as sa
-from alembic import op
-
-import TEMPLATE.models as models
-
-# revision identifiers, used by Alembic.
-revision = "6bfebb4af6ba"
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "grpc_status",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("job_hash", sa.String(), nullable=True),
-        sa.Column("job_request", sa.String(), nullable=True),
-        sa.Column("status", sa.Enum(models.Status), nullable=True),
-        sa.Column("timestamp", sa.DateTime(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("job_hash"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("grpc_status")
-    op.execute("DROP TYPE status")
-
-    # ### end Alembic commands ###
