--- SciXTEMPLATE/tests/stubdata/AVRO_schemas/TEMPLATEInputSchema.avsc
@@ -1,90 +1,0 @@
-{
-	"type": "record",
-	"name": "TemplateInput",
-	"fields": [
-		{
-			"name": "hash",
-			"type": [
-				"string",
-				"null"
-			],
-			"doc": "The job hash generated by the gRPC server."
-		},
-		{
-			"name": "id",
-			"type": [
-				"int",
-				"null"
-			],
-			"doc": "The job ID generated by storing it in postgres."
-		},
-		{
-			"name": "task",
-			"type": {
-				"type": "enum",
-				"name": "taskenum",
-				"symbols": [
-					"SYMBOL1",
-					"SYMBOL2",
-					"SYMBOL3",
-					"MONITOR"
-				]
-			},
-			"doc": "An enumerated type."
-		},
-		{
-			"name": "status",
-			"type": [
-				"null",
-				{
-					"type": "enum",
-					"name": "statusenum",
-					"symbols": [
-						"Pending",
-						"Processing",
-						"Error",
-						"Success"
-					]
-				}
-			],
-			"doc": "An enumerated type for specifying status.",
-			"default": null
-		},
-		{
-			"name": "task_args",
-			"type": {
-				"type": "record",
-				"name": "task_args_record",
-				"fields": [
-					{
-						"name": "ingest",
-						"type": [
-							"null",
-							"boolean"
-						]
-					},
-					{
-						"name": "ingest_type",
-						"type": [
-							"null",
-							"string"
-						]
-					},
-                    {
-						"name": "daterange",
-						"type": ["null", "string"]
-                    },
-					{
-						"name": "persistence",
-						"type": [
-							"boolean",
-							"null"
-						],
-						"default": false
-					}
-				]
-			},
-			"doc": "The arguments provided for the specific task."
-		}
-	]
-}
